visualize/requirements.txt
jpype1
pandas
streamlit
xlsxwriter